[MOD/#13] OnlyTextTopBar -> CenteredTextTopBar 수정
